TEAM_AMBITIOUS_Game_Tcha/requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.0.76
numpy==1.26.0
requests==2.31.0
pillow==10.2.0

